Added comments to Summer Precip script
Added revised intro comments for data inputs and outputs. Added a new folder variable to be used if the script is run again. There is a bookmark in PyCharm for this. Other parts of the script will need to be updated as well in places the input folder is referenced. I just changed the directory structure on my computer to separate the input and output files.